+++ frontend/src/General/Buttons/AddDeviceButton.js
