+++ frontend/src/General/Buttons/StreamButton.js
